@@ -1,6 +1,6 @@
 [package]
 authors = ["Riccardo Casatta <riccardo@casatta.it>", "Dev Random <c1.devrandom@niftybox.net>"]
-edition = "2018"
+edition = "2021"
 readme = "README.md"
 name = "embedded"
 version = "0.1.0"
@@ -17,7 +17,6 @@ panic-halt = "0.2.0"
 alloc-cortex-m = "0.4.1"
 bitcoin = { path="../", default-features = false, features = ["no-std", "secp-lowmemory"] }
 
-
 [[bin]]
 name = "embedded"
 test = false
@@ -33,3 +32,6 @@ path = "../../hashes"
 
 [patch.crates-io.bitcoin-internals]
 path = "../../internals"
+
+[patch.crates-io.bitcoin-io]
+path = "../../io"
