@@ -1,6 +1,7 @@
 [package]
 name = "bitcoin-fuzz"
-edition = "2018"
+edition = "2021"
+rust-version = "1.56.1"
 version = "0.0.1"
 authors = ["Generated by fuzz/generate-files.sh"]
 publish = false
