@@ -9,16 +9,17 @@ description = "Hash functions used by the rust-bitcoin eccosystem"
 categories = ["algorithms"]
 keywords = [ "crypto", "bitcoin", "hash", "digest" ]
 readme = "README.md"
-edition = "2018"
+edition = "2021"
+rust-version = "1.56.1"
 exclude = ["tests", "contrib"]
 
 [features]
 default = ["std"]
-std = ["alloc", "hex/std"]
+std = ["alloc", "hex/std", "bitcoin-io/std"]
 alloc = ["hex/alloc"]
 serde-std = ["serde/std"]
-# If you want I/O you must enable either "std" or "core2".
-core2 = ["actual-core2", "hex/core2"]
+# If you want I/O you must enable either "std" or "io".
+io = ["bitcoin-io"]
 # Smaller (but slower) implementation of sha256, sha512 and ripemd160
 small-hash = []
 
@@ -29,12 +30,10 @@ rustdoc-args = ["--cfg", "docsrs"]
 [dependencies]
 hex = { package = "hex-conservative", version = "0.1.1", default-features = false }
 
-schemars = { version = "0.8.3", optional = true }
+bitcoin-io = { version = "0.1", default-features = false, optional = true }
+schemars = { version = "0.8.3", default-features = false, optional = true }
 # Only enable this if you explicitly do not want to use "std", otherwise enable "serde-std".
 serde = { version = "1.0", default-features = false, optional = true }
-
-# Do NOT use this feature! Use the "core2" feature instead.
-actual-core2 = { package = "core2", version = "0.3.2", default-features = false, optional = true }
 
 [dev-dependencies]
 serde_test = "1.0"
