@@ -2,7 +2,7 @@
 name = "schemars"
 version = "0.1.0"
 authors = ["Jeremy Rubin <j@rubin.io>"]
-edition = "2018"
+edition = "2021"
 
 # Prevent this from interfering with workspaces
 [workspace]
@@ -21,3 +21,6 @@ serde_json = "1.0"
 
 [patch.crates-io.bitcoin-internals]
 path = "../../../internals"
+
+[patch.crates-io.bitcoin-io]
+path = "../../../io"
