@@ -34,10 +34,9 @@ macro_rules! impl_consensus_encoding {
             fn consensus_decode<R: $crate::io::Read + ?Sized>(
                 r: &mut R,
             ) -> Result<$thing, $crate::consensus::encode::Error> {
-                use crate::io::Read as _;
                 let mut r = r.take($crate::consensus::encode::MAX_VEC_SIZE as u64);
                 Ok($thing {
-                    $($field: $crate::consensus::Decodable::consensus_decode(r.by_ref())?),+
+                    $($field: $crate::consensus::Decodable::consensus_decode(&mut r)?),+
                 })
             }
         }
