@@ -17,8 +17,7 @@ use crate::{ripemd160, sha256, FromSliceError};
 crate::internal_macros::hash_type! {
     160,
     false,
-    "Output of the Bitcoin HASH160 hash function. (RIPEMD160(SHA256))",
-    "crate::util::json_hex_string::len_20"
+    "Output of the Bitcoin HASH160 hash function. (RIPEMD160(SHA256))"
 }
 
 type HashEngine = sha256::HashEngine;
