@@ -17,8 +17,7 @@ use crate::{hex, sha256d, FromSliceError, HashEngine as _};
 crate::internal_macros::hash_type! {
     256,
     false,
-    "Output of the SHA256 hash function.",
-    "crate::util::json_hex_string::len_32"
+    "Output of the SHA256 hash function."
 }
 
 #[cfg(not(hashes_fuzz))]
