@@ -12,8 +12,7 @@ use crate::{sha256, FromSliceError};
 crate::internal_macros::hash_type! {
     256,
     true,
-    "Output of the SHA256d hash function.",
-    "crate::util::json_hex_string::len_32"
+    "Output of the SHA256d hash function."
 }
 
 type HashEngine = sha256::HashEngine;
